Represent the console and client IDs as int64_t

+++ netplay/client/client.h
@@ -17,8 +17,8 @@ class NetplayClientInterface {
   virtual bool PlugControllers(const std::vector<Port>& ports,
                                PlugControllerResponsePB::Status* status) = 0;
   virtual int delay_frames() const = 0;
-  virtual int console_id() const = 0;
-  virtual int client_id() const = 0;
+  virtual int64_t console_id() const = 0;
+  virtual int64_t client_id() const = 0;
   virtual const std::vector<Port>& local_ports() const = 0;
 
   virtual TimingsPB* mutable_timings() = 0;
@@ -47,7 +47,7 @@ class NetplayClient : public NetplayClientInterface<ButtonsType> {
   // Creates a new client with the given local frame delay and console ID.
   NetplayClient(std::shared_ptr<NetPlayServerService::StubInterface> stub,
                 std::unique_ptr<ButtonCoderInterface<ButtonsType>> coder,
-                int delay_frames, int console_id);
+                int delay_frames, int64_t console_id);
 
   // Request that the given ports be plugged into the server's virtual console.
   // Returns the resulting status code returned from the server for this
@@ -57,8 +57,8 @@ class NetplayClient : public NetplayClientInterface<ButtonsType> {
 
   // Accessors
   int delay_frames() const override { return delay_frames_; }
-  int console_id() const override { return console_id_; }
-  int client_id() const override { return client_id_; }
+  int64_t console_id() const override { return console_id_; }
+  int64_t client_id() const override { return client_id_; }
   const std::vector<Port>& local_ports() const override { return local_ports_; }
   TimingsPB* mutable_timings() override { return &timings_; }
 
@@ -70,10 +70,10 @@ class NetplayClient : public NetplayClientInterface<ButtonsType> {
 
  private:
   const int delay_frames_;
-  const int console_id_;
+  const int64_t console_id_;
   std::unique_ptr<ButtonCoderInterface<ButtonsType>> coder_;
   // Client ID, set by the PlugControllers method.
-  int client_id_;
+  int64_t client_id_;
 
   std::vector<Port> local_ports_;
